+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-04-28
+
+### Added
+
+- Added [Stalled-stream protection](https://github.com/awslabs/aws-sdk-rust/discussions/956) support(enabled by default).
+- Added Express One Zone integration tests.
+
+### Changed
+
+- aws-sdk-s3 = "1.24.0"
+
 ## [1.2.0] - 2024-03-28
 
 ### Added